refactor: dev seeder merge util in service
